Fixed win32 problem for machines that have never seen a serial port. Added extra message in autoscan mode to say "now running your program"
win32 problem discovered by Martin O'Hanlon, if a win32 machine
is brand new and has never seen a serial port before, an exception
is thrown because the SERIALCOMM registry branch does not exist.

We now just catch this exception and go ahead anyway returning
an empty port list. The next time a serial port is connected,
windows creates this registry branch and it works fine.

Also increased the driver load and unload time delays from 2 secs
to 4 secs, as on some slow machines it takes a while for windows
to load/unload drivers and we were getting spurious results
sometimes on old PCs.

The running program problem was reported by Sukkin Pang.
If you don't use findPort.py but just run your program, it scans
for a serial port and then when it has found it, it runs your program.
If there is a bug in your program, you don't know whether the scan
process has failed, or if your program has failed.

Now, first time through the auto port scan, it displays a message
saying "now running your program..."